+++ packages/lumx-vue/src/components/switch/Stories/SwitchDefault.vue
@@ -1,0 +1,26 @@
+<template>
+    <Switch v-bind="attrs" :is-checked="isChecked" @change="handleChange" />
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue';
+import { Switch } from '@lumx/vue';
+import { useAttrsWithoutHandlers } from '@lumx/vue/stories/utils/useAttrsWithoutHandlers';
+
+const props = defineProps<{
+    isChecked?: boolean;
+    onChange?: (isChecked: boolean, value?: string, name?: string, event?: Event) => any;
+}>();
+
+defineOptions({
+    inheritAttrs: false,
+});
+
+const attrs = useAttrsWithoutHandlers();
+const isChecked = ref(props.isChecked || false);
+
+const handleChange = (newChecked: boolean, value?: string, name?: string, event?: Event) => {
+    isChecked.value = newChecked;
+    props.onChange?.(newChecked, value, name, event);
+};
+</script>