+++ packages/lumx-vue/src/components/switch/index.ts
@@ -1,0 +1,3 @@
+export { default as Switch } from './Switch';
+export type { SwitchProps } from './Switch';
+export { CLASSNAME, COMPONENT_NAME, DEFAULT_PROPS } from './Switch';